Make convert test fixtures less sloppy
There were some compile errors and other odds and ends that I didn't find
when I was initially writing these. The whole point of this exercise was to
enable other people to write clean, clear tests with Ginkgo, so there's no
value in writing the tests for this feature in such a slapdash fashion.