Covermode flag (and reworked pass-through flags passing) (#281)
* Rework Go arg passing

This consolidates Go flags, limiting the number of things passed into
the test runner. It also removes the intermediary definitions of the
various flags from RunWatchAndBuildCommandFlags and TestRunner, leaving
them only in the flags() and CompileTo() functions.

This is prep for adding yet another Go option.

* Adding covermode flag

To allow for a count-mode coverage run.