Include captured output from failed tests into JUnit
JUnit has the ability to report the full output of a test run via the
system-out field which is not currently exposed. This commit turns the
test Writer in a capturing writer even when streaming so that the JUnit
reporter will have access to CapturedOutput from the Spec. Failed tests
include that output.

For suites that generate a lot of log output, being able to zoom in to a
particular failed test and see the output all at once makes various
JUnit visualization outputs more effective (like in Jenkins).