Don't colorize output by default if not writing to a TTY
The current behavior of `ginkgo` is to write ANSI escape codes to the
output stream to colorize messages by default. This leads to mangled and
unreadable logs unless the user knows to pass `--noColor`. Instead, the
default reporter should check to see if the output file descriptor is a
TTY and colorize output only if it is, or if the user has specifically
provides the `--color` flag.

This patch breaks backward compatibility for users that expected ANSI
escape codes printed to non-TTY file descriptors, but those users can
pass `--color` to return to the previous behavior. All other users
should not notice this change.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>