Pass `--noColor` to test binary for compatibility
If a user compiles their test binary with an older version of the Ginkgo
library, but uses a newer version of the Ginkgo test runner to run it,
they will not be able to pass the new flag `--color` to the test binary.
Instead, we should continue pasing the `--noColor` flag to preserve
compatibility.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>