Print help to stdout when the user asks for help
This changes behaviour of `ginkgo -h`. Before this commit, it used to
print a different usage from what `ginkgo help` prints. This is caused
by go's inbuilt flag library choosing the print the `defaultUsage` when
the `Usage` func is not defined on a `flagSet`.

Another thing to note is `ginkgo -h` exits with exit code 2, `ginkgo
help` exits with 0. This is consistent with previous behaviour. Ideally
we should change `ginkgo -h` to exit with 0, but it is currently
hard-coded in the flag library.

PS: `ginkgo --help` behaves exactly the same as `ginkgo -h`.

[Fixes #567]