Defer running top-level container nodes until RunSpecs is called
A common lifecycle gotcha is to try to dynamically generate tests based on configuration data.  This fix defers evaluation of the top-level container blocks (Describe, Context, etc.) until RunSpecs.  This allows the user to load any configuration information in the TestX hook just before RunSpecs is called and leverage that configuration information to inform how the testing tree is generated.

[Fixes #693]