[v2] First Public Release of V2
Ginkgo V2 is an extensive effort to clean up the Ginkgo codebase and
introduce several new highly-requested features.

This first public release represents a significant clean up of the
codebase, deprecation of several pieces of functionality that wil not
be making it into V2, and the beginnings of some new functionality.

Details of changes so far are in the `docs/MIGRATING_TO_V2.md` file.

More extensive discussion of V2 is in #711 and the google doc linked
there.