chore: update go.sum to fix Travis CI
In CI we run `go mod tidy` and check that it hasn't changed `go.sum` in
order to check that we haven't comitted an "untidy" file. It seems that
Go 1.14 (in development) does some extra tidying of `go.sum` which is
backwards compatible with Go 1.13 and Go 1.12. Because of this backwards
compatability, it feels safe to commit this. If changes are made in
future that are not backwards compatible, then we should just run this
check on the lowest compatible version of Go.