reporters/junit: Use `system-out` element instead of `passed` (#769)
In #586, the JUnit reporter was extended to capture the output of passing tests
when the -reportPassing flag is used.

However, the passed element is not in the JUnit XML schema. While that alone
might a theoretical problem, it is also a practical problem, because some
parsers, e.g., the TeamCity XML Reporting Plugin JUnit parser, conform to that
schema, ignore the element, and show no output for passing tests.

The fix seems to be to use the system-out element instead. I have confirmed
that, when using the system-out attribute, the TeamCity XML Reporting Plugin
Junit parser shows output for a passing test.