Implement best-effort OutputInterceptor for windows
The unix OutputInterceptor leverages sycall.Dup2 to shunt all output bound for Stdout and Stderr over to a buffer.  This syscall does not exist on windows.  Instead we set os.Stdout and os.Stderr to pipe writers on windows in the hope that most code emitting output is using those globals.

This commit also changes the unix implementation to use an in-memory pipe as opposed to a temporary file.  It looks like this has won some small performance gains.