Add timestamp field to JUnitTestCase
The scope of PR is to add the TimeStamp field for JUnitTestCase.
The JUnit XML format does not have the timestamp as an attribute.

This field will give us the start time for each test. We intend to
use it to group all the tests that faill with all the tests that
run in the same time.