Ginkgo supports Interruptible Nodes and Timeouts
Nodes can now be made interruptible by passing in functions that take a SpecContext. Interruptible nodes can be governed by a deadline specified by the NodeTimeout and SpecTimeout decorators.  When the deadline expires the context is cancelled and the node must exit before a Grace Period elapses.  This Grace Period can be specified globally via the --grace-period cli flag or on a per-node basis with the GracePeriod decorator.

Moreover, interrupts are handled similarly.  When an interrupt signal is received the currently running node is interrupted by cancelling its SpecContext.  The node is then givne a Grace Period to clean up before execution continues with subsequent cleanup nodes.