Override environment formatter colors from envvars
This allows a user to set environment variables
`GINKGO_CLI_COLOR_<color>` with asci code overrides. This helps a lot
when the colorscheme that you are using doesn't play well with the
colors from the 256 color scheme which currently is hard coded.

Example `GINKGO_CLI_COLOR_RED=$'\x1b[31m' ginkgo` will use the red from
the 8bit colors configured in the terminal.