ignore hidden files
Hidden files are defined here as those beginning with '.' or '_'.

This is a departure from previous behavior, where hidden source and test files
were evaluated via watch. Evaluating hidden files is undesirable because some
editors create hidden temporary files, and these files shouldn't trigger test
runs.

No flag is added to restore the original behavior at this time as the new
behavior is the same as "go test", and it is not expected that it will cause
problems. If this change affects you, consider opening a PR to add a CLI flag
to restore the previous behavior.