Update tests
The trailing `""` to wrap empty labels was a CSV error. This commit
simply updates the test files to be properly encoded. It is a result
of running the updated `ginkgo` binary from the previous commit across
all of the _testdata files.

----

There was prior art within `outline/outline.go` which tried to
properly enclose a comma-separated list of labels within double
quotes. This resulted in CSV output with trailing `""` characters for
tests that have no labels.

A trailing `""` is incorrect for CSV formatting.

There are two special characters within CSV formatting:

    "  0x22   DOUBLE QUOTE
    ,  0x2c   COMMA

The mechanism to embed the field-separator 0x2c within a string is to
enclose it within quotation marks. The mechanism to embed the
quotation mark within a string is to use two quotation marks. i.e.
the following string:

    // go code
    foo := "The dog looked at me and said, \"Bark, bark!\""

should be represented in CSV as

    Punchline,Laughs
    "The dog looked at me and said, ""Bark, bark!""",17

just as it would be represnted in JSON as

    {
        "punchline": "The dog looked at me and said, \"Bark, bark!\"",
        "laughs": 17
    }