support transforming node parameters
This enables frameworks built on top of Ginkgo to intercept all parameters
passed to DSL functions and check and/or modify the text and arguments.

Previously it was necessary to convert users of such a framework to invoke
Ginkgo only through separate wrapper functions for e.g. ginkgo.It, which
led to churn in the test source code and confusion when arguments only
understood by a framework were passed directly to Ginkgo.