add CurrentTreeConstructionNodeReport
When using wrappers for Ginkgo functions or the ginkgo.TransformNodeArgs hook,
then code which gets invoked during the tree construction may want to know
where it is being called. The new CurrentTreeConstructionNodeReport provides
similar information to GetCurrentSpec.

This will be used in Kubernetes to add the traditional tags like "[Slow]"
if "Slow" was set as label for the current leaf node or any of its parents.