ginkgo: report exit result in case of failure
When a worker unexpectedly dies, it's hard to determine why. Even if output is
available (--output-interceptor-mode=none), if it gets killed by the Linux OOM
killer there's nothing in the process output about that.

Including a textual description of the exit status provides that
information (here simulated with `killall -KILL e2e.test`):

    Output from proc 1:
        I1204 10:03:40.282670  196272 e2e.go:109] Starting e2e run "82689063-1787-40c3-be7d-c4677c556063" on Ginkgo node 1

    Exit result of proc 1:
      signal: killed